MP_django_dev/requirements.txt
gunicorn
django
djangorestframework
whitenoise

numpy
pandas
bokeh
scipy
networkx
jgraph
wordcloud
emoji
gensim
hdbscan
nltk
numba